+++ src/LegoBalance/__init__.py
@@ -47,4 +47,4 @@ __all__ = [
     "StateEstimator",
 ]
 
-__version__ = "1.0.2"
+__version__ = "1.0.3"